hal: Fix constructor signature and remove rtapi_app's access to HAL internals.
